+++ geomesa-index-api/src/main/scala/org/locationtech/geomesa/index/utils/KryoLazyDensityUtils.scala
@@ -9,18 +9,12 @@
 package org.locationtech.geomesa.index.utils
 
 import com.vividsolutions.jts.geom.{Coordinate, Envelope, Geometry, Point}
-import org.apache.commons.codec.binary.Base64
-import org.geotools.factory.Hints
 import org.geotools.filter.text.ecql.ECQL
 import org.geotools.util.Converters
-import org.locationtech.geomesa.features.ScalaSimpleFeature
 import org.locationtech.geomesa.features.SerializationOption.SerializationOptions
 import org.locationtech.geomesa.features.kryo.KryoFeatureSerializer
-import org.locationtech.geomesa.index.conf.QueryHints.{ENCODE_STATS, RichHints, STATS_STRING}
-import org.locationtech.geomesa.utils.collection.CloseableIterator
-import org.locationtech.geomesa.utils.geotools.{GeometryUtils, GridSnap}
+import org.locationtech.geomesa.utils.geotools.GridSnap
 import org.locationtech.geomesa.utils.interop.SimpleFeatureTypes
-import org.locationtech.geomesa.utils.stats.{Stat, StatSerializer}
 import org.opengis.feature.simple.{SimpleFeature, SimpleFeatureType}
 import org.opengis.filter.expression.Expression
 
@@ -181,66 +175,5 @@ object KryoLazyDensityUtils {
         (x, y, weight)
       }
     }
-  }
-}
-
-object KryoLazyStatsUtils {
-  val StatsSft = SimpleFeatureTypes.createType("stats:stats", "stats:String,geom:Geometry")
-
-  /**
-    * Encodes a stat as a base64 string.
-    *
-    * Creates a new serializer each time, so don't call repeatedly.
-    *
-    * @param stat stat to encode
-    * @param sft simple feature type of underlying schema
-    * @return base64 string
-    */
-  def encodeStat(stat: Stat, sft: SimpleFeatureType): String =
-    Base64.encodeBase64URLSafeString(StatSerializer(sft).serialize(stat))
-
-  /**
-    * Decodes a stat string from a result simple feature.
-    *
-    * Creates a new serializer each time, so not used internally.
-    *
-    * @param encoded encoded string
-    * @param sft simple feature type of the underlying schema
-    * @return stat
-    */
-  def decodeStat(encoded: String, sft: SimpleFeatureType): Stat =
-    StatSerializer(sft).deserialize(Base64.decodeBase64(encoded))
-
-  /**
-    * Reduces computed simple features which contain stat information into one on the client
-    *
-    * @param features iterator of features received per tablet server from query
-    * @param hints query hints that the stats are being run against
-    * @return aggregated iterator of features
-    */
-  def reduceFeatures(sft: SimpleFeatureType, hints: Hints)
-                    (features: CloseableIterator[SimpleFeature]): CloseableIterator[SimpleFeature] = {
-    val transform = hints.getTransform
-    val serializer = transform match {
-      case Some((tdef, tsft)) => StatSerializer(tsft)
-      case None               => StatSerializer(sft)
-    }
-
-    val decodedStats = features.map { f =>
-      serializer.deserialize(Base64.decodeBase64(f.getAttribute(0).toString))
-    }
-
-    val sum = if (decodedStats.isEmpty) {
-      // create empty stat based on the original input so that we always return something
-      Stat(sft, hints.get(STATS_STRING).asInstanceOf[String])
-    } else {
-      val sum = decodedStats.next()
-      decodedStats.foreach(sum += _)
-      sum
-    }
-    decodedStats.close()
-
-    val stats = if (hints.containsKey(ENCODE_STATS) && hints.get(ENCODE_STATS).asInstanceOf[Boolean]) encodeStat(sum, sft) else sum.toJson
-    Iterator(new ScalaSimpleFeature("stat", StatsSft, Array(stats, GeometryUtils.zeroPoint)))
   }
 }
